requirements.txt
Django==2.1.3
django-filter==2.2.0
django-filters==0.2.1
django-rest-framework==0.1.0
django-suit==0.2.26
djangorestframework==3.10.2
psycopg2==2.8.3
pytz==2019.2
